requirements.txt
# Document Summarizer Bot Requirements
# =====================================
# Installation: pip install -r requirements.txt
# Python Version: 3.8+

# Core Dependencies
openai
requests
python-dotenv
tiktoken

# Document Processing
PyPDF2
python-docx
pdf2image

# OCR and Image Processing
pytesseract
Pillow

# Language Detection and Translation
langdetect
deep-translator

# NLP and Text Processing
nltk
scikit-learn

# Progress & UI
tqdm
ttkbootstrap

# Optional APIs
anthropic

<<<<<<< HEAD
# Language Detection
langdetect

# Translation
deep-translator

# Text Processing
scikit-learn
nltk

# Optional: Alternative PDF Processing (if PyPDF2 fails)
# pdfplumber>=0.10.0

# Optional: OCR for scanned PDFs
# pytesseract>=0.3.10
# Pillow>=10.0.0

# Optional: Advanced text processing
# spacy>=3.7.0

# Optional: Environment variables management
# python-dotenv>=1.0.0

# GUI Support (included with Python)
# tkinter (usually pre-installed with Python)

ttkbootstrap
PyPDF2
python-docx
requests
langdetect
deep-translator
python-dotenv
pytesseract
pdf2image
Pillow
=======
# Optional / Advanced
# pdfplumber>=0.10.0  # Alternative PDF Processing if PyPDF2 fails
# spacy>=3.7.0        # Advanced text processing
# tkinter              # GUI support (usually pre-installed with Python)
>>>>>>> cbc12445e53e80bc34583bc73c77bd79edd438e4
